[BugFix]: Fix DeepSeek V4 HiCache layer count logic (#25477)
